backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.63.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==25.0.3
httplib2==0.31.1
idna==3.11
numpy==1.24.3
packaging==26.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydotplus==2.0.2
pyparsing==3.3.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
